Added Facebook Authentication
The Facebook authentication API is unsupported when run within the
Unity editor or as a standalone app. Therefore, the Facebook
authentication API only works once the project is built and run on a
device or within the Facebook Canvas. For debugging, there is a “Login”
button that does not use the Facebook authentication API, but rather
grants access without requiring authentication.

Added ability to view players in same room

Added logout ability